clean up unused code and fixed weather data bug
