+++ client/src/App.js
@@ -9,8 +9,15 @@ import Login from "./pages/Login";
 import LogoutPage from "./pages/LogoutPage";
 import LoginContext from "./utils/LoginContext";
 import Navbar from "./components/Navbar/Navbar";
+import WeatherContext from "./utils/WeatherContext";
 
 function App() {
+  const [weather, setWeather] = useState({
+    temp: "",
+    windspeed: "",
+    icon: "",
+  });
+
   const [isLoggedIn, setIsLoggedIn] = useState({
     isSignedIn: false,
     checkIfLoggedIn: (response) => {
@@ -30,7 +37,6 @@ function App() {
       });
     }
 
-
     return setIsLoggedIn({
       ...isLoggedIn,
       isSignedIn: true,
@@ -43,14 +49,16 @@ function App() {
     <div className="bg-green-500">
       <Router>
         <LoginContext.Provider value={{ isLoggedIn, setIsLoggedIn }}>
-           <Navbar />
-          <Route exact path="/" component={Login} />
-          <Switch>  
-            <PrivateRoute path="/home" component={Home} />
-            <PrivateRoute path="/calendar" component={Calendar} />
-            <PrivateRoute path="/successqr" component={SuccessQR} />
-          </Switch>
-          <Route path="/logout" component={LogoutPage} />
+          <WeatherContext.Provider value={{ weather, setWeather }}>
+            <Navbar />
+            <Route exact path="/" component={Login} />
+            <Switch>
+              <PrivateRoute path="/home" component={Home} />
+              <PrivateRoute path="/calendar" component={Calendar} />
+              <PrivateRoute path="/successqr" component={SuccessQR} />
+            </Switch>
+            <Route path="/logout" component={LogoutPage} />
+          </WeatherContext.Provider>
         </LoginContext.Provider>
       </Router>
     </div>
