--- client/src/components/Footer/Footer.js
@@ -1,12 +1,0 @@
-/* eslint-disable no-unused-vars */
-import React from "react";
-
-function Footer() {
-  return (
-    <div>
-      <p>Hello Footer!</p>
-    </div>
-  );
-}
-
-export default Footer;
