--- client/src/components/Wrapper/Wrapper.js
@@ -1,12 +1,0 @@
-/* eslint-disable no-unused-vars */
-import React from "react";
-
-function Wrapper() {
-  return (
-    <div>
-      <p>Hello wrapper!</p>
-    </div>
-  );
-}
-
-export default Wrapper;
