--- client/src/pages/AvailWash.js
@@ -1,14 +1,0 @@
-// // **This page is for future development and is intentionally commented out. 
-
-// /* eslint-disable no-unused-vars */
-// import React from "react";
-
-// function AvailWash() {
-//   return (
-//     <div>
-//       <p>Hello Available Washes page!</p>
-//     </div>
-//   );
-// }
-
-// export default AvailWash;
