+++ client/src/utils/WeatherContext.js
@@ -1,0 +1,5 @@
+import React, { createContext } from "react";
+
+const WeatherContext = createContext();
+
+export default WeatherContext;
