--- models/Appointment.js
@@ -1,22 +1,0 @@
-const mongoose = require("mongoose");
-const Schema = mongoose.Schema;
-
-const AppointmentSchema = new Schema({
-  title: {
-    type: String,
-    required: true,
-  },
-  start: {
-    type: Date,
-    required: true,
-  },
-  end: {
-    type: Date,
-    required: true,
-  }
-});
-
-const Appointment = mongoose.model("appointments", AppointmentSchema);
-
-
-module.exports = Appointment;
