+++ models/index.js
@@ -1,5 +1,4 @@
 module.exports = {
   User: require("./User"),
-  Appointment: require("./Appointment"),
-  Event: require("./Event")
+  Event: require("./Event"),
 };
